Address review comments and fix tests.
