+++ estimagic/optimization/utilities.py
@@ -103,15 +103,6 @@ def dimension_to_number_of_triangular_elements(dim):
     return int(dim * (dim + 1) / 2)
 
 
-def index_element_to_string(element, separator="_"):
-    if isinstance(element, (tuple, list)):
-        as_strings = [str(entry).replace("-", "_") for entry in element]
-        res_string = separator.join(as_strings)
-    else:
-        res_string = str(element)
-    return res_string
-
-
 def propose_algorithms(requested_algo, algos, number=3):
     """Propose a a number of algorithms based on similarity to the requested algorithm.
